microbiome analysis/lib/kmer_pipeline/requirements.txt
# Note: if modifying these requirements, please also update the requirements in install/create_conda_envs.sh
numpy==1.14.3
biom_format==2.1.6
scipy==1.0.0
matplotlib==2.1.1
Keras==2.1.3
seaborn==0.8.1
scikit_learn==0.19.1
tqdm==4.19.5
biopython==1.71
